requirements.txt
requests==2.32.3
gcsa
python-dotenv

